Remove Segger RTT from pico-btstack (#1995)
* Remove btstack Segger RTT from pico-btstack

We now have this in the SDK so adding it again in pico-btstack leads to
link errors.

But it's possible to use the btstack RTT functionality by defining
ENABLE_SEGGER_RTT=1, and if we're not using pico_stio_rtt we still need
to add the btstack RTT source.

* Fix build issues when RTT is enabled

Fixes https://github.com/raspberrypi/pico-examples/issues/565

+++ src/rp2_common/pico_stdio_rtt/stdio_rtt.c
@@ -26,11 +26,11 @@ void stdio_rtt_deinit(void) {
 }
 
 static void stdio_rtt_out_chars(const char *buf, int length) {
-    SEGGER_RTT_Write(0, buf, length);
+    SEGGER_RTT_Write(0, buf, (unsigned)length);
 }
 
 static int stdio_rtt_in_chars(char *buf, int length) {
-    return SEGGER_RTT_Read(0, buf, length);
+    return (int)SEGGER_RTT_Read(0, buf, (unsigned)length);
 }
 
 stdio_driver_t stdio_rtt = {
